Refactor PersonTracking feature to improve organization and add configuration classes
