+++ src/Opilo.HazardZoneMonitor.Api/Features/Floors/GetFloors/GetFloorsResponse.cs
@@ -1,3 +1,4 @@
+using Opilo.HazardZoneMonitor.Api.Features.Floors.Configuration;
 using Opilo.HazardZoneMonitor.Api.Shared.Cqrs;
 
 namespace Opilo.HazardZoneMonitor.Api.Features.Floors.GetFloors;
