+++ src/Opilo.HazardZoneMonitor.Api/Features/Floors/GetFloors/Handler.cs
@@ -1,5 +1,6 @@
 using Ardalis.Result;
 using Microsoft.Extensions.Options;
+using Opilo.HazardZoneMonitor.Api.Features.Floors.Configuration;
 using Opilo.HazardZoneMonitor.Api.Shared.Cqrs;
 
 namespace Opilo.HazardZoneMonitor.Api.Features.Floors.GetFloors;
