+++ src/Opilo.HazardZoneMonitor.Api/Features/HazardZones/Configuration/HazardZoneConfiguration.cs
@@ -1,6 +1,6 @@
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 
-namespace Opilo.HazardZoneMonitor.Api.Features.HazardZones;
+namespace Opilo.HazardZoneMonitor.Api.Features.HazardZones.Configuration;
 
 public sealed record HazardZoneConfiguration(
     HazardZoneName Name,
