+++ src/Opilo.HazardZoneMonitor.Api/Features/HazardZones/GetHazardZones/Feature.cs
@@ -2,9 +2,11 @@ using System.ComponentModel;
 using Ardalis.Result;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.Extensions.Options;
-using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
+using Opilo.HazardZoneMonitor.Api.Features.HazardZones.Configuration;
 using Opilo.HazardZoneMonitor.Api.Shared.Cqrs;
 using Opilo.HazardZoneMonitor.Api.Shared.Features;
+using Opilo.HazardZoneMonitor.Api.Shared.Infrastructure;
+using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 
 namespace Opilo.HazardZoneMonitor.Api.Features.HazardZones.GetHazardZones;
