+++ src/Opilo.HazardZoneMonitor.Api/Features/HazardZones/GetHazardZones/GetHazardZonesResponse.cs
@@ -1,3 +1,4 @@
+using Opilo.HazardZoneMonitor.Api.Features.HazardZones.Configuration;
 using Opilo.HazardZoneMonitor.Api.Shared.Cqrs;
 
 namespace Opilo.HazardZoneMonitor.Api.Features.HazardZones.GetHazardZones;
