+++ src/Opilo.HazardZoneMonitor.Api/Features/PersonTracking/Data/IMovementsRepository.cs
@@ -1,7 +1,8 @@
 using Ardalis.Result;
+using Opilo.HazardZoneMonitor.Api.Features.PersonTracking.GetRegisteredPersonMovement;
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 
-namespace Opilo.HazardZoneMonitor.Api.Features.PersonTracking;
+namespace Opilo.HazardZoneMonitor.Api.Features.PersonTracking.Data;
 
 public interface IMovementsRepository
 {
