+++ src/Opilo.HazardZoneMonitor.Api/Features/PersonTracking/RegisterPersonMovement/Command.cs
@@ -1,3 +1,4 @@
+using Opilo.HazardZoneMonitor.Api.Features.PersonTracking.GetRegisteredPersonMovement;
 using Opilo.HazardZoneMonitor.Api.Shared.Cqrs;
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
