+++ src/Opilo.HazardZoneMonitor.Api/Features/Site/Configuration/SiteConfiguration.cs
@@ -1,0 +1,5 @@
+using Opilo.HazardZoneMonitor.Api.Features.Floors.Configuration;
+
+namespace Opilo.HazardZoneMonitor.Api.Features.Site.Configuration;
+
+public sealed record SiteConfiguration(string Name, IReadOnlyList<FloorConfiguration> Floors);
