+++ src/Opilo.HazardZoneMonitor.Api/Features/Site/Configuration/SiteOptions.cs
@@ -1,6 +1,6 @@
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 
-namespace Opilo.HazardZoneMonitor.Api.Features.Site;
+namespace Opilo.HazardZoneMonitor.Api.Features.Site.Configuration;
 
 public sealed record SiteOptions
 {
