+++ src/Opilo.HazardZoneMonitor.Api/Features/Site/GetSite/GetSiteResponse.cs
@@ -1,3 +1,4 @@
+using Opilo.HazardZoneMonitor.Api.Features.Site.Configuration;
 using Opilo.HazardZoneMonitor.Api.Shared.Cqrs;
 
 namespace Opilo.HazardZoneMonitor.Api.Features.Site.GetSite;
