--- src/Opilo.HazardZoneMonitor.Api/Features/Site/SiteConfiguration.cs
@@ -1,5 +1,0 @@
-using Opilo.HazardZoneMonitor.Api.Features.Floors;
-
-namespace Opilo.HazardZoneMonitor.Api.Features.Site;
-
-public sealed record SiteConfiguration(string Name, IReadOnlyList<FloorConfiguration> Floors);
