+++ src/Opilo.HazardZoneMonitor.Api/Shared/Database/CoordinateTypeHandler.cs
@@ -3,7 +3,7 @@ using System.Text.Json;
 using Dapper;
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 
-namespace Opilo.HazardZoneMonitor.Api.Features.PersonTracking;
+namespace Opilo.HazardZoneMonitor.Api.Shared.Database;
 
 internal sealed class CoordinateTypeHandler : SqlMapper.TypeHandler<Coordinate>
 {
