+++ src/Opilo.HazardZoneMonitor.Api/Shared/Database/GuidTypeHandler.cs
@@ -1,7 +1,7 @@
 using System.Data;
 using Dapper;
 
-namespace Opilo.HazardZoneMonitor.Api.Features.PersonTracking;
+namespace Opilo.HazardZoneMonitor.Api.Shared.Database;
 
 internal sealed class GuidTypeHandler : SqlMapper.TypeHandler<Guid>
 {
