+++ tests/Opilo.HazardZoneMonitor.Api.Tests.Unit/Features/Floors/FloorOptionsValidatorTests.cs
@@ -1,4 +1,4 @@
-using Opilo.HazardZoneMonitor.Api.Features.Floors;
+using Opilo.HazardZoneMonitor.Api.Features.Floors.Configuration;
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 using Opilo.HazardZoneMonitor.Tests.Common.TestUtilities.Builders;
