+++ tests/Opilo.HazardZoneMonitor.Api.Tests.Unit/Features/HazardZones/GetHazardZones/HandlerTests.cs
@@ -1,6 +1,6 @@
 using Ardalis.Result;
 using Microsoft.Extensions.Options;
-using Opilo.HazardZoneMonitor.Api.Features.HazardZones;
+using Opilo.HazardZoneMonitor.Api.Features.HazardZones.Configuration;
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 using Opilo.HazardZoneMonitor.Api.Features.HazardZones.GetHazardZones;
