+++ tests/Opilo.HazardZoneMonitor.Api.Tests.Unit/Features/HazardZones/HazardZoneOptionsValidatorTests.cs
@@ -1,4 +1,4 @@
-using Opilo.HazardZoneMonitor.Api.Features.HazardZones;
+using Opilo.HazardZoneMonitor.Api.Features.HazardZones.Configuration;
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 using Opilo.HazardZoneMonitor.Tests.Common.TestUtilities.Builders;
