+++ tests/Opilo.HazardZoneMonitor.Tests.Common/TestUtilities/Builders/HazardZoneConfigurationBuilder.cs
@@ -1,4 +1,4 @@
-using Opilo.HazardZoneMonitor.Api.Features.HazardZones;
+using Opilo.HazardZoneMonitor.Api.Features.HazardZones.Configuration;
 using Opilo.HazardZoneMonitor.Domain.Shared.Primitives;
 
 namespace Opilo.HazardZoneMonitor.Tests.Common.TestUtilities.Builders;
